Use `click_clack/` Not `src/` Package for Compatibility & Support More Option Types
This will be another release 0.1.1. Should make this tool more broadly usable in more codebases, in more situations. Previously, the fact that this was packaged in the `src/` directory meant that pip was for some reason clobbering the `src/` dir of any project that tried installing this package (meaning that that project's own `src/` dir was broken and couldn't be referenced via `from src import foo` - which is absolutely wild but apparently how Python works?.... absolutely broken ass language). Now I'm only claiming the `click_clack/` dir, which realistically shouldn't be an issue for users.